+++ test/abci_test.go
@@ -2696,3 +2696,7 @@ func TestDisableAllNode(t *testing.T) {
 		DisableNode(t, param)
 	}
 }
+
+func TestGetChainHistory(t *testing.T) {
+	GetChainHistory(t)
+}