+++ CHANGELOG.md
@@ -7,6 +7,10 @@ IMPROVEMENTS:
 - [DeliverTx] Add `chain_history_info` property to parameters of `InitNDID`.
 - [Query] Add new function (`GetChainHistory`).
 
+SECURIT FIXES:
+
+- [CheckTx] Check duplicate `nonce` in every transaction.
+
 ## 0.11.2 (November 12, 2018)
 
 IMPROVEMENTS: