+++ CHANGELOG.md
@@ -7,9 +7,9 @@ IMPROVEMENTS:
 - [DeliverTx] Add `chain_history_info` property to parameters of `InitNDID`.
 - [Query] Add new function (`GetChainHistory`).
 
-SECURIT FIXES:
+SECURITY FIXES:
 
-- [CheckTx] Check duplicate `nonce` in every transaction.
+- [CheckTx] Check duplicate `nonce` in every transactions.
 
 ## 0.11.2 (November 12, 2018)